whoa: refactor provider to deal with network errors
