+++ kinode/default_providers_testnet.json
@@ -1,76 +1,78 @@
 [
     {
-        "chain_id": 1,
-        "trusted": false,
-        "provider": {
-            "RpcUrl": "wss://ethereum.publicnode.com"
-        },
-        "public": false,
-        "allow": [],
-        "deny": []
-    },
-    {
         "chain_id": 11155111,
-        "trusted": false,
+        "trusted": true,
         "provider": {
-            "RpcUrl": "wss://eth-sepolia.g.alchemy.com/v2/already_deleted_it"
-        },
-        "public": false,
-        "allow": [],
-        "deny": []
+            "Node": {
+                "use_as_provider": true,
+                "kns_update": {
+                    "name": "sepoliarocks.os",
+                    "owner": "",
+                    "node": "0x2b2e9479333c5f94b62a242d75298ce98d13ad0af95070bc0b8d35aacdbddfa7",
+                    "public_key": "0x958a3f43aee848826db2c0b36545e1e775bf310b003f0d7abf72ab8697a1b72c",
+                    "ip": "",
+                    "port": 0,
+                    "routers": [
+                        "default-router-1.os",
+                        "default-router-2.os",
+                        "default-router-3.os"
+                    ]
+                }
+            }
+        }
     },
     {
         "chain_id": 11155111,
         "trusted": true,
         "provider": {
             "Node": {
-                "name": "default-router-1.os",
-                "owner": "",
-                "node": "0xb35eb347deb896bc3fb6132a07fca1601f83462385ed11e835c24c33ba4ef73d",
-                "public_key": "0xb1b1cf23c89f651aac3e5fd4decb04aa177ab0ec8ce5f1d3877b90bb6f5779db",
-                "ip": "147.135.114.167",
-                "port": 9002,
-                "routers": []
+                "use_as_provider": false,
+                "kns_update": {
+                    "name": "default-router-1.os",
+                    "owner": "",
+                    "node": "0xb35eb347deb896bc3fb6132a07fca1601f83462385ed11e835c24c33ba4ef73d",
+                    "public_key": "0xb1b1cf23c89f651aac3e5fd4decb04aa177ab0ec8ce5f1d3877b90bb6f5779db",
+                    "ip": "147.135.114.167",
+                    "port": 9002,
+                    "routers": []
+                }
             }
-        },
-        "public": false,
-        "allow": [],
-        "deny": []
+        }
     },
     {
         "chain_id": 11155111,
         "trusted": true,
         "provider": {
             "Node": {
-                "name": "default-router-2.os",
-                "owner": "",
-                "node": "0xd827ae579fafa604af79fbed977e8abe048497f10885c6473dfd343a3b7b4458",
-                "public_key": "0xab9f1a996db3a4e1dbcd31d765daedeb3af9af4f570c0968463b5be3a7d1e992",
-                "ip": "147.135.114.167",
-                "port": 9003,
-                "routers": []
+                "use_as_provider": false,
+                "kns_update": {
+                    "name": "default-router-2.os",
+                    "owner": "",
+                    "node": "0xd827ae579fafa604af79fbed977e8abe048497f10885c6473dfd343a3b7b4458",
+                    "public_key": "0xab9f1a996db3a4e1dbcd31d765daedeb3af9af4f570c0968463b5be3a7d1e992",
+                    "ip": "147.135.114.167",
+                    "port": 9003,
+                    "routers": []
+                }
             }
-        },
-        "public": false,
-        "allow": [],
-        "deny": []
+        }
     },
     {
         "chain_id": 11155111,
         "trusted": true,
         "provider": {
             "Node": {
-                "name": "default-router-3.os",
-                "owner": "",
-                "node": "0x96e36331c8f0882f2c0c46c13b15d812def04fe8606d503bc0e2be39db26486a",
-                "public_key": "0x536e30785e64dd0349a697285af365b5ed7c4d300010139261cfc4dbdd5b254b",
-                "ip": "147.135.114.167",
-                "port": 9004,
-                "routers": []
+                "use_as_provider": false,
+                "kns_update": {
+                    "name": "default-router-3.os",
+                    "owner": "",
+                    "node": "0x96e36331c8f0882f2c0c46c13b15d812def04fe8606d503bc0e2be39db26486a",
+                    "public_key": "0x536e30785e64dd0349a697285af365b5ed7c4d300010139261cfc4dbdd5b254b",
+                    "ip": "147.135.114.167",
+                    "port": 9004,
+                    "routers": []
+                }
             }
-        },
-        "public": false,
-        "allow": [],
-        "deny": []
+        }
     }
 ]
