+++ kinode/packages/app_store/pkg/manifest.json
@@ -13,6 +13,12 @@
             "vfs:distro:sys",
             "kernel:distro:sys",
             "eth:distro:sys",
+            {
+                "process": "eth:distro:sys",
+                "params": {
+                    "root": true
+                }
+            },
             "sqlite:distro:sys",
             "kv:distro:sys",
             "chess:chess:sys",
